requirements.txt
firebase-admin
fastapi
bson
uvicorn
python-dotenv
pymongo